Faster, Better, Stronger
Removed the silly nested for loop and imported larger list of English words from nltk.corpus. Code now runs through the list of English words and checks to see if the word matches the criteria instead of brute forcing a bunch of made up words that meet the criteria and comparing them to real words.